cleaning up statement property component
