+++ app/src/app/components/stix/name-property/name-property.component.scss
@@ -8,6 +8,10 @@
             text-transform: none;
             top: 0 !important;
             margin-right: 1ex;
+        
+        }
+        .mat-form-field-wrapper {
+            padding-bottom: 0;
         }
     }
 }
