--- app/src/app/components/stix/statement-property/statement-diff/statement-diff.component.html
@@ -1,1 +1,0 @@
-<p>statement-diff works!</p>
