--- app/src/app/components/stix/statement-property/statement-diff/statement-diff.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-statement-diff',
-  templateUrl: './statement-diff.component.html',
-  styleUrls: ['./statement-diff.component.scss']
-})
-export class StatementDiffComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
