--- app/src/app/components/stix/statement-property/statement-edit/statement-edit.component.scss
@@ -1,11 +1,0 @@
-@import "../../../../../style/globals";
-
-.statement-edit {
-    button {
-        &:hover {
-            cursor: pointer;
-            .dark &  { border-color: on-color-deemphasis(dark) !important; }
-            .light & { border-color: on-color-deemphasis(light) !important; }
-        }
-    }
-}
