+++ app/src/app/components/stix/statement-property/statement-property.component.html
@@ -1,7 +1,44 @@
 <div class="statement-property">
     <div class="content">
-        <app-statement-diff *ngIf="config.mode == 'diff'" [config]="config"></app-statement-diff>
-        <app-statement-view *ngIf="config.mode == 'view'" [config]="config"></app-statement-view>
-        <app-statement-edit *ngIf="config.mode == 'edit'" [config]="config"></app-statement-edit>
+        <div *ngIf="(data$ | async)">
+            <div [ngClass]="config.mode == 'edit' ? 'statement-edit' : ''" *ngIf="objStatements.length > 0 else noStatements" >
+                <button *ngIf="config.mode == 'view'" popoverPlacement="bottom" 
+                    [popover]="objectStatus"
+                    [popoverOnHover]="true"
+                    [disabled]="disabled"
+                    >
+                    <div *ngIf="objStatements.length == 1">1 statement</div>
+                    <div *ngIf="objStatements.length > 1">{{objStatements.length}} statements</div>
+                </button>
+                <button *ngIf="config.mode == 'edit'" popoverPlacement="bottom" 
+                    [popover]="objectStatus"
+                    [popoverOnHover]="true"
+                    [disabled]="disabled"
+                    (click)="updateStatements()"
+                    >
+                    <div *ngIf="objStatements.length == 1">1 statement</div>
+                    <div *ngIf="objStatements.length > 1">{{objStatements.length}} statements</div>
+                </button>
+            </div>
+        
+            <popover-content #objectStatus [closeOnClickOutside]="true">
+                <div>
+                    <center>Statements</center>
+                    <ul>
+                        <li *ngFor="let objStatement of objStatements">{{objStatement.definition_string}}</li>
+                    </ul>
+                    <center *ngIf="config.mode == 'edit'">Click to add more statements</center>
+                </div>
+            </popover-content>
+        
+            <ng-template #noStatements>
+                <div [ngClass]="config.mode == 'edit' ? 'statement-edit' : ''">
+                    <button *ngIf="config.mode == 'view'">
+                        0 statements
+                    </button>
+                    <button *ngIf="config.mode == 'edit'" (click)="updateStatements()">+ add statement</button>
+                </div>
+            </ng-template>
+        </div>
     </div>
 </div>
