+++ app/src/app/components/stix/statement-property/statement-property.component.scss
@@ -1,0 +1,11 @@
+@import "../../../../style/globals";
+
+.statement-edit {
+    button {
+        &:hover {
+            cursor: pointer;
+            .dark &  { border-color: on-color-deemphasis(dark) !important; }
+            .light & { border-color: on-color-deemphasis(light) !important; }
+        }
+    }
+}
