--- app/src/app/components/stix/statement-property/statement-view/statement-view.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { StatementViewComponent } from './statement-view.component';
-
-describe('StatementViewComponent', () => {
-  let component: StatementViewComponent;
-  let fixture: ComponentFixture<StatementViewComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ StatementViewComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(StatementViewComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
